Move towards manifests/ charts instead of install/kubernetes/helm/ charts (#385)
* Adapt to CRD changes

* Implement base resource changes

* Implement gateways changes
