+++ pkg/resources/base/rbac.go
@@ -14,9 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package common
+package base
 
 import (
+	"fmt"
+
 	apiv1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -26,8 +28,9 @@ import (
 func (r *Reconciler) serviceAccount() runtime.Object {
 	return &apiv1.ServiceAccount{
 		ObjectMeta: v1.ObjectMeta{
-			Name:      istioReaderName,
+			Name:      istioReaderServiceAccountName,
 			Namespace: r.Config.Namespace,
+			Labels:    istioReaderLabel,
 		},
 	}
 }
@@ -35,9 +38,15 @@ func (r *Reconciler) serviceAccount() runtime.Object {
 func (r *Reconciler) clusterRole() runtime.Object {
 	return &rbacv1.ClusterRole{
 		ObjectMeta: v1.ObjectMeta{
-			Name: istioReaderName,
+			Name:   r.istioReaderNameWithNamespace(),
+			Labels: istioReaderLabel,
 		},
 		Rules: []rbacv1.PolicyRule{
+			{
+				APIGroups: []string{"config.istio.io", "rbac.istio.io", "security.istio.io", "networking.istio.io", "authentication.istio.io"},
+				Resources: []string{"*"},
+				Verbs:     []string{"get", "watch", "list"},
+			},
 			{
 				APIGroups: []string{""},
 				Resources: []string{"nodes", "pods", "services", "endpoints", "replicationcontrollers"},
@@ -55,19 +64,24 @@ func (r *Reconciler) clusterRole() runtime.Object {
 func (r *Reconciler) clusterRoleBinding() runtime.Object {
 	return &rbacv1.ClusterRoleBinding{
 		ObjectMeta: v1.ObjectMeta{
-			Name: istioReaderName,
+			Name:   r.istioReaderNameWithNamespace(),
+			Labels: istioReaderLabel,
 		},
 		RoleRef: rbacv1.RoleRef{
 			Kind:     "ClusterRole",
 			APIGroup: "rbac.authorization.k8s.io",
-			Name:     istioReaderName,
+			Name:     r.istioReaderNameWithNamespace(),
 		},
 		Subjects: []rbacv1.Subject{
 			{
 				Kind:      "ServiceAccount",
-				Name:      istioReaderName,
+				Name:      istioReaderServiceAccountName,
 				Namespace: r.Config.Namespace,
 			},
 		},
 	}
+}
+
+func (r *Reconciler) istioReaderNameWithNamespace() string {
+	return fmt.Sprintf("%s-%s", istioReaderName, r.Config.Namespace)
 }
